pyproject.toml
[project]
name = "virustotal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alive-progress>=3.3.0",
    "argparse>=1.4.0",
    "asyncio>=4.0.0",
    "configparser>=7.2.0",
    "google-genai>=1.61.0",
    "logging>=0.4.9.6",
    "psutil>=7.2.2",
    "requests>=2.32.5",
]
